# services/bank-svc/swagger.yaml
openapi: 3.0.0
info:
  title: Bank Service
  description: APIs that manage financial institution work
  version: 1.0.0

paths:
  /v1/bank/accounts:
    post:
      summary: Save information about a new bank account
      operationId: saveBankAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveNewAccountRequest'
      responses: 
        '200':
          description: OK; Saved
        '400':
          description: Invalid request
        '500':
          description: Server error
  /v1/bank/institutions:
    post:
      summary: Save Financial Institution Information
      operationId: saveNewBank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveNewBankRequest'
      responses:
        '201':
          description: Financial Institution successfully saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveNewBankResponse'
        '400':
          description: Invalid request
        '500':
          description: Server error
    get:
      summary: Get all Financial Institutions
      operationId: getFinancialInstitutions
      responses:
        '200':
          description: Currently registered financial institutions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialInstitutionsResponse'
        '400':
          description: Invalid request
        '500':
          description: Server Error
  /v1/bank/users/{userID}/banks:
    get:
      summary: Get a user's financial institution information
      operationId: getUserFinancialProfile
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetched user's bank details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFinancialProfile'
        '400':
          description: Invalid request
        '500':
          description: Server error

components:
  schemas:
    SaveNewBankRequest:
      type: object
      required:
        - bank_name
      properties:
        bank_name:
          type: string
    SaveNewBankResponse:
      type: object
      required:
        - bank_id
      properties:
        bank_id:
          type: integer

    SaveNewAccountRequest:
      type: object
      required:
        - user_id
        - bank_name
        - accounts
        - access_token
      properties:
        user_id:
          type: integer
        bank_name:
          type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        access_token:
          type: string

    FinancialInstitutionsResponse:
      type: object
      required:
        - banks
      properties:
        banks:
          type: array
          items:
            $ref: '#/components/schemas/BankItem'
    BankItem:
      type: object
      required:
        - bank_id
        - bank_name
      properties:
        bank_id:
          type: integer
        bank_name:
          type: string

    UserFinancialProfile:
      type: object
      required:
        - banks
      properties:
        user_id:
          type: integer
        banks:
          type: array
          items:
            $ref: '#/components/schemas/UserAccountInfo'
    
    UserAccountInfo:
      type: object
      required:
        - bank_name
        - accounts
        - access_token
      properties:
        bank_name:
          type: string
        access_token:
          type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'

    Account:
      type: object
      required:
        - account_id
        - account_type
      properties:
        account_id:
          type: integer
        account_type:
          type: string